requirements.txt
allure-pytest==2.13.0
allure-python-commons==2.13.0
Appium-Python-Client==2.7.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
data==0.4
decorator==5.1.1
distlib==0.3.4
exceptiongroup==1.1.0
execnet==1.9.0
Faker==13.6.0
filelock==3.6.0
Flask==2.1.3
funcsigs==1.0.2
greenlet==2.0.2
h11==0.13.0
idna==3.3
importlib-metadata==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.1
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
urllib3-secure-extra==0.1.0
virtualenv==20.13.4
webdriver-manager==3.8.5
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.11.0